Add privacy policy page
This commit adds a new `privacy-policy.html` page to the repository. 
The page contains an initial version of the privacy policy for our website, 
detailing how we collect, use, and share user data. 

Sections included are:
- Introduction
- Data Collection
- Data Use
- Data Sharing
- Contact Information

This is the first draft of the privacy policy, and further revisions may follow.